backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Werkzeug==3.0.1
gunicorn==21.2.0
pymongo==4.6.1
PyJWT==2.8.0
bcrypt==4.1.2
Flask-Limiter==3.5.0
requests==2.32.4
numpy==2.3.2
python-dotenv==1.1.1
openai==1.6.1
chromadb==0.4.22
